% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mtg_pals.R
\docType{data}
\name{mtg_colors}
\alias{mtg_colors}
\title{this code is heavily based on
https://drsimonj.svbtle.com/creating-corporate-colour-palettes-for-ggplot2}
\format{An object of class \code{character} of length 7.}
\usage{
mtg_colors
}
\description{
this code is heavily based on
https://drsimonj.svbtle.com/creating-corporate-colour-palettes-for-ggplot2
}
\keyword{datasets}
