% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mtg_pals.R
\name{mtg_cols}
\alias{mtg_cols}
\title{Function to extract MtG colors as hex codes}
\usage{
mtg_cols(...)
}
\arguments{
\item{...}{Character names of mtg_colors. Options are "White", "Blue",
"Black",  "Red", "Green", "Multicolored", and "Colorless".}
}
\description{
Function to extract MtG colors as hex codes
}
